Some big, not all pleasant, changes to PersistentObjects.
IPersistentObject and its children are now covariant in TEntity.
IPersistentObject no longer has a GetEntity() method, because Task<TEntity> is not covariant in TEntity.
ICharacter has a GetCharacterEntity() method with a fixed CharacterEntity return type. ShardSession needs access to character snapshots for a variety of reasons.
BuildUpdate methods in IPersistentObject now take ObjectEntity as an argument, so they can be used to build creation updates for other objects. I would like to revert this change soon, as it is not optimal, but it does prevent a particular type of message cycle.
Added ObjectService as a way to view children of PersistentObject as PersistentObjects.